+++ app/Http/Controllers/CustomerController.php
@@ -77,6 +77,12 @@ class CustomerController extends Controller
      */
     public function delete($id)
     {
+        $res = $this->microgen->service('customers')->deleteById($id);
+
+        if (array_key_exists('error', $res)) {
+            return response()->json($res['error'], $res['status']);
+        };
+
         Customer::where('id', '=', $id)->delete();
 
         return response()-> json(array("message" => "Resource deleted successfully"));
